requirements.txt
feedparser
requests
python-dotenv
boto3
